Ten findings, and two of them were the same mistake
The roll control was keyed on one ChunkRef, which was wrong in two ways
at once. A pack may attach several roll tables to a chunk -- nothing in
the schema makes tables.chunk_id unique -- and toMap silently kept
whichever the database returned last, so one control was unreachable and
the surviving one was a function of iteration order. And when a nested
rollable table is deduplicated into its parent quote, the table's text
ends up on screen while the capability stays keyed to the child, so the
card contained a table and offered no way to roll on it. Card.Verbatim
carries the chunks it can roll on, Survivor records what it absorbed,
and the app renders one labelled control per table.

The full activation gate ran on every question. Library.openActive
reached the same validation the CLI's direct-file path needs -- scanning
every chunk's text, decoding every vector -- for packs that install had
already gated, the digest pass re-hashes, and borrow had open-checked
moments earlier. That was the direct-file rule applied to a case it was
not written for.

The app persisted Kotlin's data-class toString. AskService's default
render is cards.joinToString, so a feed restored after process death read
Verbatim(ref=ChunkRef(...)) -- and that text was destined to become the
context a follow-up resolves against. Answer.asPlainText keeps the quote
gutter, because a stored answer that cannot say which lines were the
book's own words fails the one guarantee this product makes, in storage
rather than on screen.

The rest:

- The builder scored a summary that decomposes to no claim as 1.0
  without asking the judge anything. ClaimReport.vacuous already
  answered this; the builder had its own copy of the arithmetic and not
  of the rule.
- distribution() summed outcomes in Int while min, max and every table
  bound are Long, so d6+2147483647 produced wrapped negative keys and an
  exact-PMF API disagreed with the roller about the same expression.
- Capability loading materialized every chunk's kind and origin to
  classify targets that may not exist -- on every question, for a pack
  that may have no capabilities at all.
- Supersession impact counted the pack being replaced, so a new edition
  correcting its own predecessor measured as broad, installed itself
  inactive, and then deleted the copy it was measured against.
- The generation budget charged the body and not the heading path, which
  travels into the same prompt and has the same multi-megabyte ceiling.
- The refusal card discarded which packs were searched and whether
  inactive ones existed, which is the difference between a real absence
  and a book the user forgot to activate.

585 tests, and the APK assembles.

+++ pack/src/main/kotlin/dev/rpghelper/pack/DiceExpression.kt
@@ -41,10 +41,14 @@ data class DiceExpression(
      * distinguishable by equality rather than by a confidence interval: they share the
      * range 2–12 and differ in every interior probability.
      */
-    fun distribution(): Map<Int, Rational> {
-        var dist = mapOf(0 to Rational.ONE)
+    fun distribution(): Map<Long, Rational> {
+        // Long throughout, because `min`, `max` and every table row bound already are. The
+        // grammar admits a modifier up to `Int.MAX_VALUE`, so `d6+2147483647` summed in Int
+        // produced wrapped negative keys -- an exact-PMF API disagreeing with the roller and
+        // the coverage logic about what the very same expression can produce.
+        var dist = mapOf(0L to Rational.ONE)
         repeat(count) {
-            val next = HashMap<Int, Rational>()
+            val next = HashMap<Long, Rational>()
             for ((total, probability) in dist) {
                 for (face in 1..sides) {
                     val outcome = total + face
